fix: decorator types (#1685)
I found this issue https://github.com/willsoto/nestjs-prometheus/issues/1667 that had the same errors as me, but the comments didn't help. I also found a different issue https://github.com/nestjs/nest/issues/10959, and the fix they used https://github.com/nestjs/nest/pull/10970, and used that to fix the problem with the InjectMetric decorator.

The problem is that the decorator types are wrong. Record<string, unknown> fails with strict checks, and the key may also be undefined.

Fixes #1667